backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.135.1",
    "google-genai>=1.66.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.48",
    "uvicorn[standard]>=0.41.0",
]
